Fix search button
