+++ src/components/molecules/GeoJsonForm.vue
@@ -62,7 +62,6 @@ export default defineComponent({
       };
       return true;
     };
-
     return {
       useMap,
       clearShape,
